move to vault.go
